moved mnist to models
