+++ tests/models/mnist.py
